refactor(setting): 重构设置页面视图模型结构
- 将前台UI自定义功能拆分为独立的partial类文件
- 移除已迁移的自动更新、语言设置和快捷入口相关代码
- 保留调试选项区域的原有功能
- 优化代码组织结构提升可维护性
